rest-api/requirements.txt
Flask
flask_login
flask_sqlalchemy
psycopg2-binary
SQLAlchemy-Utils
python-dotenv
Flask-Cors
